+++ create/config.py
@@ -1,0 +1,1 @@
+default_host = 'localhost'
